Saving and Loading implemented
Populations can now be saved on a whim and loaded back in via NQ_Save
and NQ_Load. NQ_Start and NQ_End also support parameter which will call
the respective file I/O.

+++ src/gene.h
@@ -32,7 +32,7 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 //Type definitions for our neural network.
 typedef struct gene_s
 {
-	double innovation_num; // Used for finding corresponding genes during crossover.
+	int innovation_num; // Used for finding corresponding genes during crossover.
 	double mutation_num; // Used to see how much mutation has changed the link.
 
 	cbool enabled; // Disables the gene.
